restructure draw commands into a draw module
