+++ arcade/__init__.py
@@ -110,37 +110,37 @@ from .texture import make_soft_square_texture
 from .texture import get_default_image
 from .texture import get_default_texture
 
-from .draw_commands import get_points_for_thick_line
-from .draw_commands import draw_arc_filled
-from .draw_commands import draw_arc_outline
-from .draw_commands import draw_circle_filled
-from .draw_commands import draw_circle_outline
-from .draw_commands import draw_ellipse_filled
-from .draw_commands import draw_ellipse_outline
-from .draw_commands import draw_line
-from .draw_commands import draw_line_strip
-from .draw_commands import draw_lines
-from .draw_commands import draw_lrbt_rectangle_filled
-from .draw_commands import draw_lrbt_rectangle_outline
-from .draw_commands import draw_lbwh_rectangle_textured
-from .draw_commands import draw_parabola_filled
-from .draw_commands import draw_parabola_outline
-from .draw_commands import draw_point
-from .draw_commands import draw_points
-from .draw_commands import draw_polygon_filled
-from .draw_commands import draw_polygon_outline
-from .draw_commands import draw_rect_filled
-from .draw_commands import draw_rect_outline
-from .draw_commands import draw_scaled_texture_rectangle
-from .draw_commands import draw_texture_rectangle
-from .draw_commands import draw_triangle_filled
-from .draw_commands import draw_triangle_outline
-from .draw_commands import draw_lbwh_rectangle_filled
-from .draw_commands import draw_lbwh_rectangle_outline
-from .draw_commands import draw_rect_filled_kwargs
-from .draw_commands import draw_rect_outline_kwargs
-from .draw_commands import get_image
-from .draw_commands import get_pixel
+from .draw import get_points_for_thick_line
+from .draw import draw_arc_filled
+from .draw import draw_arc_outline
+from .draw import draw_circle_filled
+from .draw import draw_circle_outline
+from .draw import draw_ellipse_filled
+from .draw import draw_ellipse_outline
+from .draw import draw_line
+from .draw import draw_line_strip
+from .draw import draw_lines
+from .draw import draw_lrbt_rectangle_filled
+from .draw import draw_lrbt_rectangle_outline
+from .draw import draw_lbwh_rectangle_textured
+from .draw import draw_parabola_filled
+from .draw import draw_parabola_outline
+from .draw import draw_point
+from .draw import draw_points
+from .draw import draw_polygon_filled
+from .draw import draw_polygon_outline
+from .draw import draw_rect_filled
+from .draw import draw_rect_outline
+from .draw import draw_scaled_texture_rectangle
+from .draw import draw_texture_rectangle
+from .draw import draw_triangle_filled
+from .draw import draw_triangle_outline
+from .draw import draw_lbwh_rectangle_filled
+from .draw import draw_lbwh_rectangle_outline
+from .draw import draw_rect_filled_kwargs
+from .draw import draw_rect_outline_kwargs
+from .draw import get_image
+from .draw import get_pixel
 
 # We don't have joysticks game controllers in headless mode
 if not pyglet.options["headless"]:
